Move FeedViewAdapter to separate file
